requirements.txt
adlfs==2022.11.2
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.13.1
anyio==3.7.1
apprise==1.7.1
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
azure-core==1.29.6
azure-datalake-store==0.0.53
azure-identity==1.15.0
azure-storage-blob==12.19.0
bcrypt==4.1.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coolname==2.2.0
croniter==2.0.1
cryptography==41.0.7
dacite==1.8.1
dateparser==1.2.0
dgl==0.9.1
dnspython==2.4.2
docker==5.0.3
email-validator==2.1.0.post1
exceptiongroup==1.2.0
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.26.2
gqlalchemy==1.4.0
graphviz==0.20.1
greenlet==3.0.3
griffe==0.38.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
isodate==0.6.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kubernetes==28.1.0
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msal==1.26.0
msal-extensions==1.1.0
multidict==6.0.4
neo4j==4.4.11
networkx==2.8.8
numpy==1.26.3
oauthlib==3.2.2
orjson==3.9.10
packaging==23.2
paramiko==3.4.0
pathspec==0.12.1
pendulum==2.1.2
portalocker==2.8.2
prefect==2.14.15
psutil==5.9.7
pyarrow==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
PyJWT==2.8.0
pymgclient==1.3.1
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python-telegram-bot==20.7
pytz==2023.3.post1
pytzdata==2020.1
pywin32==227
PyYAML==6.0.1
readchar==4.0.5
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
telegram==0.0.1
telepot==12.7
text-unidecode==1.3
toml==0.10.2
torch==1.13.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.25.0
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
zipp==3.17.0